add oci generative ai service
