--- api/core/model_runtime/model_providers/oci_genai/oci_genai.yaml
@@ -1,34 +1,0 @@
-provider: oci_genai
-label:
-  en_US: OCIGenerativeAI
-description:
-  en_US: Models provided by OCI, such as Cohere Command R and Cohere Command R+.
-  zh_Hans: OCI 提供的模型，例如 Cohere Command R 和 Cohere Command R+。
-icon_small:
-  en_US: icon_s_en.png
-icon_large:
-  en_US: icon_l_en.png
-background: "#FFFFFF"
-help:
-  title:
-    en_US: Get your API Key from MistralAI
-    zh_Hans: 从 OCI 获取 API Key
-  url:
-    en_US: https://cloud.oracle.com
-supported_model_types:
-  - llm
-  #- text-embedding
-  #- rerank
-configurate_methods:
-  - predefined-model
-  #- customizable-model
-provider_credential_schema:
-  credential_form_schemas:
-    - variable: google_api_key
-      label:
-        en_US: API Key
-      type: secret-input
-      required: true
-      placeholder:
-        zh_Hans: 在此输入您的 API Key
-        en_US: Enter your API Key
